backend/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.1.7
django-annoying==0.10.6
django-cors-headers==3.7.0
django-phone-field==1.8.1
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.2
future==0.18.2
html5lib==1.1
idna==2.10
imgkit==1.2.2
numpy==1.20.2
pdfkit==0.6.1
Pillow==8.2.0
plotly==4.14.3
psutil==5.8.0
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.4
WeasyPrint==52.5
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.5